Added socks5 proxy to avoid VPN
Added socks5 proxy to work remotely even with active VPN tunnel - now we can build the images
locally with connected VPN, the thing is - it's using socket.SO_DONTROUTE which skipping the route
table and just directly sends the packets to the required interfaces.

But this feature to work with VMWare packer provider needs the latest packer (not issued 1.7.9) from master.