Added project example image
The change are about adding the example new project image. It contains a different layout for the
xcode installation, contains 2 required xcode versions (12.2 and 12.4) and based on MacOS 11.06.
For sure it's better to install small tools via Jenkins Global Tools interface, but in case it's
not possible you can bake them into the image as shown in this example.

The main purpose of using 11.06 instead of 10.15 is to use Apple paravirt driver for GPU to provide
Metal support for the ios simulator tests.

The additional useful changes:
* Tested Monterey 12.3 with Xcode 13.3 for parts of the image
* Added VNCRecorder script to replace poorly usable Screenshoter - now it records the whole session
  as video file
* A bit improved the proxy script
* Split the vncdo script for vmtools installation for different macos systems (10,11,12)