Moved to use autologin on MacOS for jenkins agent
The change contains some low-level improvements and migration to the autologin for macos systems
due to the multiple reasons of instability caused by headless MacOS execution. I think it will help
with the main issue we found with unit tests stuck which most probably caused by the inflation of
WindowServer.

Changes:
* Updated the documentation and changed the build/pack script names
* Added the upload script to automate the image placing in artifact storage
* Added performance improvements to the base image and disabled auto downloading of the updates
* Improved stability of building the CI images - now they are using iso image to access macos
  recovery instead of buggy hacks in bios/efi shell
* Switched MacOS CI to autologin by default to address functionality & stability issues