Added OpenSSH for Windows and AWS init scripts for keys (#21)
Adding OpenSSH to Windows actually allows us to unify the shell access to machines and helps with transferring files to the remote instance (due to winrm issue with binary data transfers). Unfortunately packer don't want to decrypt password if we're not using winrm so I also added scripts which helps to create administrator authorized_keys file with instance pubkey in it.