Native image builder implementation (#19)
The native image is just a tar archive which can be combined with the other archives when needed. The archives are generated on the host machine (which is not great actually, because could mess with internal files - for sure it will be better to use some sort of chroot but need to unify for mac and lin).

The overall build system works as usual and needs build, pack & upload of the tar.xz archive. The naming quite similar to Docker and the image dir besides manifest & checksum file contains tar archive with files. Ansible executes locally on the host and I put a little optimization - if it runs locally download will not use tmp dir to store artifacts, but just will use files in playbooks directly.