Separation of disk management and warmup script (#57)
Found that's it's best to warmup EBS on AWS even on the dynamic environments. The important detail to do that multithreaded (to utilize bandwidth of S3) and just for a limited actually used space. For that we use FIO on all 3 platforms.

* Separated disk management from jenkins agent
* Added a few new roles to keep things organized
* Kept the same logic - but automount is installed as a part of init_ci role instead of jenkins one
* Added `diskmount_disks_warmup` option to enable warmup where it's needed

MacOS as usual gave me headache: with internal disks it works just fine, but for external disks it needs a special treatment, because SIP will not give access for the root user. To workaround that I used ssh localhost access to root used via dynamically created key: ssh entry point gets full disk access, which makes interaction with the disks much easier. This way works well for local services which are not that dependent on the user environment (like jenkins agent for example). It's not perfect by any means, but so far the best way I found.