Fixes for jenkins agent and xcode simulators (#60)
Found that `xcrun xdevice list` shows just the host simruntime available if it was not listed before even after installing the required simruntimes - so we just need to list (and validate) the simruntimes in advance to make it available in the future calls. 

Also it has a fix for introduced docker image build issue where it is unable to setup jenkins user due to unavailability of systemd.