Simplified the process of checking jenkins is available by downloading agent (#67)
Change simplifies logic of the jenkins agent startup script to validate jenkins availability by downloading the agent from it. If we have agent - that means we can continue and from there the agent itself will check that jenkins is fine or not.

Also it modifies the colima validation to limit to 5 specs - otherwise it's taking too long.