Locate `devenv.exe` using `vswhere` (#68)
For some older versions of Visual Studio, the registry does not contain a path to devenv.exe.

The visual studio 17.4.4 installer I was working with in particular does not edit the registry. From searching online, it seems Microsoft has been a bit inconsistent across releases as to what gets set where.